fix: int32_t→int, propset.h set_width hoist, DIR→MAC_DIR, parser.h inline
Four fixes from CW8 build attempt error analysis:

1. shims/stdint.h: changed int32_t from typedef long to typedef int,
   same for uint32_t. Both are 32-bit on PPC but CW8 treats long*
   and int* as incompatible pointer types, causing 34 implicit
   conversion errors across libcss parse-property files.

2. autogenerated_propset.h: hoisted the one remaining mixed
   declaration in set_width() — uint32_t *bits declared after an
   if-block. 105 other *bits declarations were already at function
   top and fine.

3. mac_dirent.h: renamed DIR typedef to MAC_DIR. The name DIR
   collides with libhubbub's element-type.h enum constant DIR,
   causing 13 "illegal name overloading" errors. Updated the 2
   callers in file.c and utils.c.

4. dom/bindings/hubbub/parser.h: replaced the ../../../../ relative
   include (which CW8 can't resolve) with an inlined copy of the
   real parser.h content. The errors.h include inside it resolves
   via the libdom:bindings:hubbub: access path.

Linux cross-check passes clean.

+++ browser/netsurf/frontends/macos9/shims/stdint.h
@@ -39,7 +39,7 @@ typedef short              int16_t;
 #if !defined(_INT32_T) && !defined(__int32_t_defined)
 #define _INT32_T
 #define __int32_t_defined
-typedef long               int32_t;
+typedef int                int32_t;
 #endif
 
 #if !defined(_INT64_T) && !defined(__int64_t_defined)
@@ -64,7 +64,7 @@ typedef unsigned short     uint16_t;
 #if !defined(_UINT32_T) && !defined(__uint32_t_defined)
 #define _UINT32_T
 #define __uint32_t_defined
-typedef unsigned long      uint32_t;
+typedef unsigned int       uint32_t;
 #endif
 
 #if !defined(_UINT64_T) && !defined(__uint64_t_defined)
